TRIC-20<feat: add endpoints to retrieve all BonSortie and BonSortie par id>

+++ src/main/java/org/tricol/supplierchain/repository/BonSortieRepository.java
@@ -6,6 +6,6 @@ import org.tricol.supplierchain.entity.BonSortie;
 
 
 @Repository
-public interface BonSortieRepository extends JpaRepository<BonSortie, Integer> {
+public interface BonSortieRepository extends JpaRepository<BonSortie, Long> {
 
 }
